+++ changelog.md
@@ -1,5 +1,8 @@
 # Change-Log
 
+## 2013-06-0 9v0.4.4
+Add isMaxX/isMinX/isMaxY/isMinY as an interface to the final TOP and bottoms positions.
+
 ## 2013-06-08 v0.4.3
 Now supports X and Y coordinates of scroller