+++ changelog.md
@@ -1,6 +1,9 @@
 # Change-Log
 
-## 2013-06-0 9v0.4.4
+## 2013-08-27 v0.4.6
+Fixed horizontal scroll (css)
+
+## 2013-06-09v0.4.4
 Add isMaxX/isMinX/isMaxY/isMinY as an interface to the final TOP and bottoms positions.
 
 ## 2013-06-08 v0.4.3